Date Format promotion update 1

+++ src/main/java/com/project/JewelryMS/service/SchedulingService.java
@@ -33,7 +33,7 @@ public class SchedulingService {
 
     // Method to assign a staff member to a shift
     @Transactional
-    public Staff_Shift assignStaffToShift(int staffId, int shiftId) {
+    public Staff_Shift assignStaffToShift(int staffId, long shiftId) {
         // Fetch the staff and shift entities from the database
         StaffAccount staff = staffAccountRepository.findById(staffId)
                 .orElseThrow(() -> new RuntimeException("Staff not found"));
@@ -127,7 +127,7 @@ public class SchedulingService {
 
     // Method to remove a staff member from a shift
     @Transactional
-    public void removeStaffFromShift(int staffId, int shiftId) {
+    public void removeStaffFromShift(int staffId, long shiftId) {
         // Fetch the staff and shift entities from the database
         StaffAccount staff = staffAccountRepository.findById(staffId)
                 .orElseThrow(() -> new RuntimeException("Staff not found"));
